# lib/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: nilsbrugger.nb.nb.nb@gmail.com
    name: Nils
  description: The time planer API
  license:
    name: Apache 2.0
  title: Time Planer
  version: 2.2.1-alpha
servers:
- url: http://eds.logfro.de/time-planer
- url: http://localhost:8080/time-planer
...
paths:
...
  /user/{username}/mygroups/:
    get:
      description: Here you can get all groups which the user owns
      operationId: getOwnedGroups
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                  - name: voluptate
                    description: labo
                    members: 20
                    creation_date: 2018-03-27T00:00:00.000Z
                    tasks: 20046389
                    uid: eVJluBrbUD
                  - name: aute ut
                    description: aliqua cu
                    members: 15
                    creation_date: 2018-03-31T00:00:00.000Z
                    tasks: -33328415
                    uid: PvjFKaSloJ
                  - name: ut adipis
                    description: in ad
                    members: 2
                    creation_date: 2018-06-15T00:00:00.000Z
                    tasks: -64953313
                    uid: IB1RTvSWOq
              schema:
                items:
                  $ref: '#/components/schemas/ReducedGroup'
                type: array
          description: The List of all Groups(reduced version) the user created
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      summary: Get all created groups
      tags:
      - Grouping
      x-accepts: application/json
      x-tags:
      - hasMore: "true"
        tag: Grouping
      - tag: MyGroups
    post:
      description: Here you can create an group
      operationId: createGroup
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitialGroup'
        x-examples:
          application/json:
            descr: 1 B Klasse der IT des TGMs
            name: 1BHIT
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                    name: nisi aliq
                    description: labore
                    members:
                    - email: bXfC0wCZ@lQOdBtLNNnAHXszqiv.zuk
                      edit: true
                      create: true
                      delete: true
                    - email: 8KUnmm4ddOQx-dk@oLmnLUlvUiDKsPaNEKkppT.ef
                      edit: true
                      create: true
                      delete: true
                    - email: 8xY9doc7NIOlj@nsTBCaeRbzyifbkp.ljn
                      edit: false
                      create: false
                      delete: true
                    - email: zONMSuT8e@fPQFIxsqZnWH.mqh
                      edit: true
                      create: false
                      delete: true
                    - email: fBIo15o2@MWgCYeCPMVvoSqwH.gbuv
                      edit: false
                      create: false
                      delete: true
                    creation_date: 1940-05-05T00:00:00.000Z
                    tasks:
                    - laboris mollit tempor cupidatat
                    - esse in
                    - dolor cupidatat sint deserunt cillum
                    - in do est eiusmod dolore
                    - eiusmod sunt fugiat
                    creator: ONDC2@bNvcKOGlvDlBpfqTpehlatAoCokiPll.rv
                    uid: qVNCy9Dv1hBF2hgfgj238gkj5jkg65
              schema:
                $ref: '#/components/schemas/Group'
          description: |-
            Group created.
            Returned the created Group
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      summary: Create a group
      tags:
      - Grouping
      x-accepts: application/json
      x-tags:
      - hasMore: "true"
        tag: Grouping
      - tag: MyGroups
      x-contentType: application/json
...
  /user/{username}/groups/:
    get:
      description: Here you can get all groups where you are a member
      operationId: getGroups
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                  - name: voluptate
                    description: labo
                    members: 20
                    creation_date: 2018-03-27T00:00:00.000Z
                    tasks: 20046389
                    uid: eVJluBrbUD
                  - name: aute ut
                    description: aliqua cu
                    members: 15
                    creation_date: 2018-03-31T00:00:00.000Z
                    tasks: -33328415
                    uid: PvjFKaSloJ
                  - name: ut adipis
                    description: in ad
                    members: 2
                    creation_date: 2018-06-15T00:00:00.000Z
                    tasks: -64953313
                    uid: IB1RTvSWOq
              schema:
                items:
                  $ref: '#/components/schemas/ReducedGroup'
                type: array
          description: The List of all Groups(reduced version) the user created
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      security:
      - User_Key: []
      summary: Get all membered groups
      tags:
      - Grouping
      x-accepts: application/json
      x-tags:
      - hasMore: "true"
        tag: Grouping
      - tag: MemberingGroups
  /user/{username}/group/{group_UUID}:
    get:
      description: Get information about a group
      operationId: getGroup
      parameters:
      - description: The users name
        explode: false
        in: path
        name: username
        required: true
        schema:
          maxLength: 20
          minLength: 3
          type: string
        style: simple
      - description: The UUID of the group to operate with
        explode: false
        in: path
        name: group_UUID
        required: true
        schema:
          maxLength: 30
          minLength: 30
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                    name: commo
                    description: ei
                    members:
                    - email: fgqLoaro@mo.ub
                      edit: true
                      create: false
                      delete: true
                    - email: fOw1H@VynIJdkJgAWihnlsVwbltPURTja.xr
                      edit: true
                      create: true
                      delete: true
                    - email: TIEpnc-4@cePagFYicdyYX.co
                      edit: false
                      create: false
                      delete: false
                    creation_date: 1966-04-05T00:00:00.000Z
                    tasks:
                    - ipsum tempor
                    - do
                    - magna
                    - non qui ad do officia
                    creator: zAQbuIqqZhWLN@JBSqByyWInaFQBdjgMSSuIuinD.ki
                    uid: GaBCJetycE
              schema:
                $ref: '#/components/schemas/ReducedGroup'
          description: The complete group in response
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key missing
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            There is no API key in the header.
            Look into the Autentication User Key to get aknowledge how it works
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            API key is wrong, get a new one.
            Maybe it expired
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: group not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: group_not_found
      security:
      - User_Key: []
      summary: Get Group Task informations
      tags:
      - Grouping
      x-accepts: application/json
      x-tags:
      - hasMore: "true"
        tag: Grouping
      - tag: MemberingGroups
components:
...
  schemas:
...
    InitialGroup:
      description: |-
        The Data Type to create an Group.

        It is the reduced form of an Group as it only contains the values needed for initialisation.
      example:
        name: 1BHIT
        description: description
      properties:
        description:
          description: A Description of the Group what it is for
          maxLength: 2500
          minLength: 0
          type: string
        name:
          description: |-
            The name of the task.

            **NOT** unique
          example: 1BHIT
          maxLength: 100
          minLength: 2
          type: string
      required:
      - description
      - name
      type: object
    Group:
      description: All information about the group
      example:
        uid: uid
        creator: creator
        members:
        - edit: true
          create: true
          delete: true
          email: email
        - edit: true
          create: true
          delete: true
          email: email
        name: 1BHIT
        description: description
        creation_date: 2000-01-23
        tasks:
        - tasks
        - tasks
      properties:
        name:
          description: The name of the group, must **NOT** be unique
          example: 1BHIT
          maxLength: 100
          minLength: 3
          type: string
        description:
          description: A short description of the Group visible for all members
          maxLength: 2500
          minLength: 0
          type: string
        members:
          description: The List of members in the Group
          items:
            $ref: '#/components/schemas/GroupMember'
          type: array
          uniqueItems: false
        creation_date:
          description: The date when the group was created
          format: date
          type: string
        tasks:
          description: The list of all tasks within the Group
          items:
            type: string
          type: array
        creator:
          description: The email of the creator of the group
          format: email
          type: string
        uid:
          description: An identifyer for the Group as multiple groups with the same name can exist
          maxLength: 30
          minLength: 30
          pattern: '[A-Z0-9a-z]{30}'
          type: string
      required:
      - creation_date
      - creator
      - description
      - members
      - name
      - tasks
      - uid
      type: object
...
    ReducedGroup:
      description: A Group only containing the complete neccessary information
      example:
        uid: uid
        members: 0
        name: 1BHIT
        description: description
        creation_date: 2000-01-23
        tasks: 5
      properties:
        name:
          description: The name of the group, must **NOT** be unique
          example: 1BHIT
          maxLength: 100
          minLength: 3
          type: string
        description:
          description: A short description of the Group visible for all members
          maxLength: 2500
          minLength: 0
          type: string
        members:
          description: The number of members in the Group
          format: int32
          minimum: 0
          type: integer
        creation_date:
          description: The date when the group was created
          format: date
          type: string
        tasks:
          description: The number of all tasks within the Group
          example: 5
          format: int32
          type: integer
        uid:
          description: An identifyer for the Group as multiple groups with the same name can exist
          maxLength: 30
          minLength: 30
          pattern: '[A-Za-z0-9]{30}'
          type: string
      required:
      - creation_date
      - description
      - members
      - name
      - tasks
      - uid
      type: object
    GroupMember:
      additionalProperties: false
      description: The Member of a group defined by his name and permissions
      example:
        edit: true
        create: true
        delete: true
        email: email
      properties:
        email:
          description: The email of the user
          format: email
          type: string
        edit:
          description: Is the user permitted to edit tasks within the group
          type: boolean
        create:
          description: is the user permitted to create tasks within the Group
          type: boolean
        delete:
          description: is the User permitted to delte tasks within the group
          type: boolean
      required:
      - create
      - delete
      - edit
      - email
      title: Member
      type: object
    ErrorResponse:
      description: Detailed information about an Error
      properties:
        message:
          description: A short description only a few words, what went wrong
          enum:
          - api key missing
          - api key wrong
          - group not found
          - group not owned
          - no password
          - no permission
          - user not found
          - params not matching pattern
          - No email address
          - wrong password
          - User allready exists
          - task_not_found
          type: string
      required:
      - message
      type: object
...
  securitySchemes:
    User_Key:
      in: header
      name: user_key
      type: apiKey